Add files via upload
The project works based on the gas sensor which will be placed in a helmet, when the sensor detects alcohol a signal is sent to the processor. When the processor gets the signal it makes sure that the bike won’t start. The LED placed will indicate the presence of alcohol.